+++ R/tm_t_mult_events.R
@@ -100,7 +100,8 @@ template_mult_events <- function(dataname,
 
   parsed_basic_table_args <- teal.widgets::parse_basic_table_args(
     teal.widgets::resolve_basic_table_args(
-      user_table = basic_table_args
+      user_table = basic_table_args,
+      module_table = teal.widgets::basic_table_args(show_colcounts = TRUE)
     )
   )
 
@@ -111,8 +112,7 @@ template_mult_events <- function(dataname,
   layout_list <- add_expr(
     layout_list,
     substitute(
-      expr = rtables::split_cols_by(var = arm_var) %>%
-        rtables::add_colcounts(),
+      expr = rtables::split_cols_by(var = arm_var),
       env = list(arm_var = arm_var)
     )
   )